requirements.txt
flask
pymysql
cryptography
python-dotenv
gunicorn